Patch kramdown vulnerability of CVE-2020-1400
An upgrade of jekyll serves the patch.  According to the commented instructions about using GitHub Pages in the file Gemfile, I have commented out the line specifying the gem version of jekyll.  Then by running `bundle install` to generate a new Gemfile.lock, the gems jekyll (and especially kramdown) will be up-to-date, getting rid of the security warning on GitHub.